Music valence prediction model
The project was part of an externship to create a music prediction valence model. there is still some fine tuning that can be done with the model, however choosing a bigger dataset would be the best path forward as it would allow better tuning. The dataset used for this is the DEAM dataset downloaded from Kaggle.